blog_api/requirements.txt
click==8.1.3
Flask==2.2.2
Flask-Pydantic==0.11.0
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pydantic==1.10.2
PyJWT==2.6.0
python-dotenv==0.21.0
SQLAlchemy==1.4.42
typing_extensions==4.4.0
Werkzeug==2.2.2
